refactor: remove in-tree wallet implementation
Keep the node-side descriptor, UTXO scan, and PSBT assembly RPCs while
removing wallet-only methods and the wallet crate boundary.
