+++ .github/workflows/ci.yml
@@ -15,13 +15,6 @@ env:
   FULL_NODE_FEATURES: "rocksdb,fjall,redb,mdbx,kernel"
 
 jobs:
-  wallet-no-seckey:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - run: |
-          ! grep -r 'SecretKey\|secp256k1::Secret\|seckey' crates/wallet/src
-
   fmt:
     runs-on: ubuntu-latest
     steps:
