+++ crates/rpc/Cargo.toml
@@ -28,7 +28,6 @@ bitcoin-rs-p2p.workspace = true
 bitcoin-rs-utxo.workspace = true
 bitcoin-rs-coinstats.workspace = true
 bitcoin-rs-mining.workspace = true
-bitcoin-rs-wallet.workspace = true
 bitcoin-rs-storage.workspace = true
 bitcoin-rs-pruning.workspace = true
 bitcoin.workspace = true
