+++ crates/rpc/README.md
@@ -1,8 +1,8 @@
 # bitcoin-rs-rpc
 
-The synchronous, Bitcoin Core-compatible JSON-RPC and REST surface of the node: method dispatch, HTTP Basic and cookie authentication, and the watch-only wallet boundary that disables every private-key RPC.
+The synchronous, Bitcoin Core-compatible JSON-RPC and REST surface of the node: method dispatch, HTTP Basic and cookie authentication, and a wallet-free method surface — every RPC that would require private key material is simply absent.
 
-`RpcServer::bind` binds a TCP listener and `serve` (or `serve_with_shutdown` for controlled shutdown) runs the blocking accept loop, handing each connection to a bounded worker thread under a per-connection idle timeout. Each request is authenticated by `Auth`, then matched to a Core-compatible handler by `Handler::dispatch`, which reads shared node state through the dependency-injected `Context` — the boundary carrying `ChainControl` consensus-affecting operations, `PruneService`, `TxIndexQuery`, `NetworkState`, and `ZmqNotification`. Failures map to JSON-RPC error codes through `RpcError`, and Bitcoin Core-compatible REST endpoints (`rest`) are served on the same listener when enabled. The surface is intentionally watch-only: RPCs that would reveal, import, create, or use private keys return an internal error, `wallet has no private keys; use external signer`, while PSBT construction, combination, analysis, and finalization remain available because they are driven by external signers without this process holding private key material.
+`RpcServer::bind` binds a TCP listener and `serve` (or `serve_with_shutdown` for controlled shutdown) runs the blocking accept loop, handing each connection to a bounded worker thread under a per-connection idle timeout. Each request is authenticated by `Auth`, then matched to a Core-compatible handler by `Handler::dispatch`, which reads shared node state through the dependency-injected `Context` — the boundary carrying `ChainControl` consensus-affecting operations, `PruneService`, `TxIndexQuery`, `NetworkState`, and `ZmqNotification`. Failures map to JSON-RPC error codes through `RpcError`, and Bitcoin Core-compatible REST endpoints (`rest`) are served on the same listener when enabled. RPCs that would reveal, import, create, or use private keys are not implemented and answer `method not found`, while PSBT combination and finalization remain available because they are driven by external signers without this process holding private key material.
 
 ## Interface architecture and implementation guidance
 
@@ -14,7 +14,7 @@ The synchronous, Bitcoin Core-compatible JSON-RPC and REST surface of the node: 
 - **JSON-RPC Framing & Protocol Versioning**: `JsonRpcVersion` (`crates/rpc/src/server.rs`) governs wire framing:
   - Requests with `"jsonrpc": "2.0"` use JSON-RPC 2.0 (`JsonRpcVersion::V2`): success responses emit `{"jsonrpc":"2.0","result":...,"id":...}` (HTTP 200), error responses emit `{"jsonrpc":"2.0","error":...,"id":...}` (HTTP 200), and requests omitting `id` are treated as notifications returning HTTP 204 No Content.
   - Other requests use JSON-RPC 1.1 / legacy (`JsonRpcVersion::Legacy`): success responses emit `{"result":...,"error":null,"id":...}` (HTTP 200), error responses emit `{"result":null,"error":...,"id":...}` with HTTP 500 status, and missing `id` values default to `null`.
-- **Watch-Only Boundary**: Methods requiring private keys return `RpcError::MethodDisabled` (`wallet has no private keys; use external signer`). PSBT endpoints remain fully functional without private key material.
+- **Wallet-Free Surface**: The node ships no wallet. Methods requiring private keys (funding, signing, key import/export) are not implemented and return `RpcError::MethodNotFound`. PSBT combination and finalization, descriptor utilities, and `scantxoutset` remain available because they are driven by external signers without this process holding private key material.
 
 ### 2. Deep module separation and system owners
 - **Adapters vs Modules**: RPC handlers, REST routes, and Esplora projections sit at the transport Seam as pure wire Adapters translating network payloads into deep Module Interfaces (`Context`, `BlockTree`, `TxIndexQuery`, `applied_tip`). They leverage domain-local concurrency and storage mechanisms without leaking database locks, indexing details, or consensus logic into protocol serialization.
