+++ crates/rpc/src/error.rs
@@ -50,7 +50,7 @@ impl RpcError {
     /// Bitcoin Core invalid parameter value code.
     pub const CORE_INVALID_PARAMETER: i64 = -8;
 
-    /// Builds the no-private-keys policy error used by signing RPCs.
+    /// Builds the policy-disabled error for methods unavailable by configuration.
     #[must_use]
     pub const fn method_disabled(message: &'static str) -> Self {
         Self::MethodDisabled(message)
