+++ crates/rpc/src/handlers/tx.rs
@@ -387,6 +387,159 @@ fn parse_txid(value: &str) -> Result<Txid, RpcError> {
     Txid::from_str(value).map_err(|_| RpcError::InvalidParams("txid must be 64 hex characters"))
 }
 
+pub(crate) fn finalizepsbt(_ctx: &Arc<Context>, params: &Value) -> Result<Value, RpcError> {
+    let raw = required_str(params, 0, "psbt is required")?;
+    let decoded = decode_base64(raw)?;
+    let Ok(psbt) = bitcoin::psbt::Psbt::deserialize(&decoded) else {
+        return Err(RpcError::InvalidParams("invalid base64 PSBT"));
+    };
+    let serialized = encode_base64(&psbt.serialize());
+    let complete = !psbt.inputs.is_empty()
+        && psbt
+            .inputs
+            .iter()
+            .all(|input| input.final_script_sig.is_some() || input.final_script_witness.is_some());
+    if complete {
+        let tx = psbt.extract_tx_unchecked_fee_rate();
+        let hex = bitcoin::consensus::encode::serialize(&tx).to_lower_hex_string();
+        Ok(json!({
+            "psbt": serialized,
+            "hex": hex,
+            "complete": true,
+        }))
+    } else {
+        Ok(json!({
+            "psbt": serialized,
+            "hex": "",
+            "complete": false,
+        }))
+    }
+}
+
+pub(crate) fn combinepsbt(_ctx: &Arc<Context>, params: &Value) -> Result<Value, RpcError> {
+    let array = params_array(params)?
+        .first()
+        .and_then(|value| value.as_array())
+        .ok_or(RpcError::InvalidParams("psbts must be an array"))?;
+    if array.is_empty() {
+        return Err(RpcError::InvalidParams("psbts array must not be empty"));
+    }
+
+    let mut iter = array.iter();
+    let Some(first_val) = iter.next() else {
+        return Err(RpcError::InvalidParams("psbts array must not be empty"));
+    };
+    let Some(first_str) = first_val.as_str() else {
+        return Err(RpcError::InvalidType("each psbt must be a string"));
+    };
+    let mut psbt = bitcoin::psbt::Psbt::deserialize(&decode_base64(first_str)?)
+        .map_err(|_| RpcError::InvalidParams("invalid base64 PSBT"))?;
+
+    for value in iter {
+        let Some(s) = value.as_str() else {
+            return Err(RpcError::InvalidType("each psbt must be a string"));
+        };
+        let other = bitcoin::psbt::Psbt::deserialize(&decode_base64(s)?)
+            .map_err(|_| RpcError::InvalidParams("invalid base64 PSBT"))?;
+        psbt.combine(other)
+            .map_err(|err| RpcError::Internal(format!("combine failed: {err}")))?;
+    }
+
+    Ok(json!(encode_base64(&psbt.serialize())))
+}
+
+const BASE64_ALPHABET: &[u8; 64] =
+    b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
+
+fn decode_base64(input: &str) -> Result<Vec<u8>, RpcError> {
+    let bytes = input.as_bytes();
+    if bytes.is_empty() || !bytes.len().is_multiple_of(4) {
+        return Err(RpcError::InvalidParams("invalid base64 PSBT"));
+    }
+
+    let chunk_count = bytes.len() / 4;
+    let mut out = Vec::with_capacity(chunk_count * 3);
+    for (index, chunk) in bytes.chunks_exact(4).enumerate() {
+        let last = index + 1 == chunk_count;
+        let pad2 = chunk[2] == b'=';
+        let pad3 = chunk[3] == b'=';
+        if chunk[0] == b'=' || chunk[1] == b'=' || pad2 && !pad3 || pad3 && !last {
+            return Err(RpcError::InvalidParams("invalid base64 PSBT"));
+        }
+
+        let Some(a) = base64_value(chunk[0]) else {
+            return Err(RpcError::InvalidParams("invalid base64 PSBT"));
+        };
+        let Some(b) = base64_value(chunk[1]) else {
+            return Err(RpcError::InvalidParams("invalid base64 PSBT"));
+        };
+        let c = if pad2 {
+            0
+        } else {
+            let Some(value) = base64_value(chunk[2]) else {
+                return Err(RpcError::InvalidParams("invalid base64 PSBT"));
+            };
+            value
+        };
+        let d = if pad3 {
+            0
+        } else {
+            let Some(value) = base64_value(chunk[3]) else {
+                return Err(RpcError::InvalidParams("invalid base64 PSBT"));
+            };
+            value
+        };
+
+        out.push((a << 2) | (b >> 4));
+        if !pad2 {
+            out.push((b << 4) | (c >> 2));
+        }
+        if !pad3 {
+            out.push((c << 6) | d);
+        }
+    }
+
+    Ok(out)
+}
+
+const fn base64_value(byte: u8) -> Option<u8> {
+    match byte {
+        b'A'..=b'Z' => Some(byte - b'A'),
+        b'a'..=b'z' => Some(byte - b'a' + 26),
+        b'0'..=b'9' => Some(byte - b'0' + 52),
+        b'+' => Some(62),
+        b'/' => Some(63),
+        _ => None,
+    }
+}
+
+fn encode_base64(bytes: &[u8]) -> String {
+    let mut out = String::with_capacity(bytes.len().div_ceil(3) * 4);
+    for chunk in bytes.chunks(3) {
+        let b0 = chunk[0];
+        let b1 = chunk.get(1).copied().unwrap_or(0);
+        let b2 = chunk.get(2).copied().unwrap_or(0);
+
+        out.push(char::from(BASE64_ALPHABET[usize::from(b0 >> 2)]));
+        out.push(char::from(
+            BASE64_ALPHABET[usize::from(((b0 & 0b0000_0011) << 4) | (b1 >> 4))],
+        ));
+        if chunk.len() > 1 {
+            out.push(char::from(
+                BASE64_ALPHABET[usize::from(((b1 & 0b0000_1111) << 2) | (b2 >> 6))],
+            ));
+        } else {
+            out.push('=');
+        }
+        if chunk.len() > 2 {
+            out.push(char::from(BASE64_ALPHABET[usize::from(b2 & 0b0011_1111)]));
+        } else {
+            out.push('=');
+        }
+    }
+    out
+}
+
 #[cfg(test)]
 mod tests {
     use alloc::sync::Arc;
@@ -1470,5 +1623,82 @@ mod gettxout_via_utxo_tests {
             value.is_null(),
             "expected null for output absent from UTXO set, got {value:?}"
         );
+    }
+}
+
+#[cfg(test)]
+mod combinepsbt_tests {
+    use alloc::sync::Arc;
+
+    use sonic_rs::JsonValueTrait as _;
+
+    use super::*;
+
+    fn empty_psbt_str() -> String {
+        let tx = bitcoin::Transaction {
+            version: bitcoin::transaction::Version(2),
+            lock_time: bitcoin::absolute::LockTime::ZERO,
+            input: Vec::new(),
+            output: Vec::new(),
+        };
+        let psbt = bitcoin::psbt::Psbt::from_unsigned_tx(tx)
+            .unwrap_or_else(|err| panic!("from_unsigned_tx: {err}"));
+        encode_base64(&psbt.serialize())
+    }
+
+    #[test]
+    fn combinepsbt_single_input_returns_same_psbt() {
+        let ctx = Arc::new(Context::new());
+        let psbt_str = empty_psbt_str();
+        let result = combinepsbt(&ctx, &json!([[psbt_str.as_str()]]))
+            .unwrap_or_else(|err| panic!("combinepsbt: {err}"));
+        let Some(out) = result.as_str() else {
+            panic!("expected string: {result:?}");
+        };
+        assert_eq!(out, psbt_str);
+    }
+
+    #[test]
+    fn combinepsbt_empty_array_errors() {
+        let ctx = Arc::new(Context::new());
+        let result = combinepsbt(&ctx, &json!([[]]));
+        assert!(result.is_err());
+    }
+}
+
+#[cfg(test)]
+mod finalizepsbt_tests {
+    use alloc::sync::Arc;
+
+    use sonic_rs::JsonValueTrait as _;
+
+    use super::*;
+
+    fn empty_psbt() -> String {
+        let tx = bitcoin::Transaction {
+            version: bitcoin::transaction::Version(2),
+            lock_time: bitcoin::absolute::LockTime::ZERO,
+            input: Vec::new(),
+            output: Vec::new(),
+        };
+        let psbt =
+            bitcoin::psbt::Psbt::from_unsigned_tx(tx).unwrap_or_else(|err| panic!("psbt: {err}"));
+        encode_base64(&psbt.serialize())
+    }
+
+    #[test]
+    fn finalizepsbt_returns_incomplete_for_unfinalized_inputs() {
+        let ctx = Arc::new(Context::new());
+        let raw = empty_psbt();
+        let result = finalizepsbt(&ctx, &json!([raw.as_str()]))
+            .unwrap_or_else(|err| panic!("finalizepsbt failed: {err}"));
+        let Some(complete) = result.get("complete").and_then(Value::as_bool) else {
+            panic!("complete missing: {result:?}");
+        };
+        assert!(!complete);
+        let Some(hex) = result.get("hex").and_then(Value::as_str) else {
+            panic!("hex missing: {result:?}");
+        };
+        assert_eq!(hex, "");
     }
 }
