+++ crates/rpc/tests/handler_smoke.rs
@@ -75,11 +75,8 @@ fn all_required_handlers_return_core_shapes() -> Result<(), Box<dyn std::error::
             "scantxoutset",
             json!(["start", ["addr(1111111111111111111114oLvT2)"]]),
         ),
-        ("walletcreatefundedpsbt", json!([[], []])),
-        ("walletprocesspsbt", json!([valid_psbt.as_str()])),
         ("finalizepsbt", json!([valid_psbt.as_str()])),
         ("combinepsbt", json!([[valid_psbt.as_str()]])),
-        ("bumpfee", json!([txid.as_str()])),
     ];
 
     for (method, params) in calls {
@@ -503,17 +500,35 @@ fn network_peer_methods_read_shared_peer_registry() -> Result<(), Box<dyn std::e
 }
 
 #[test]
-fn signing_methods_are_disabled() -> Result<(), Box<dyn std::error::Error>> {
+fn removed_wallet_methods_return_method_not_found() -> Result<(), Box<dyn std::error::Error>> {
     let handler = Handler::new(Arc::new(Context::new()));
-    let error = handler
-        .dispatch("signrawtransactionwithwallet", &json!([]))
-        .err()
-        .ok_or("signing method unexpectedly succeeded")?;
-    assert_eq!(error.code(), RpcError::INTERNAL_ERROR);
-    assert_eq!(
-        error.to_string(),
-        "wallet has no private keys; use external signer"
-    );
+    for method in [
+        "walletcreatefundedpsbt",
+        "walletprocesspsbt",
+        "bumpfee",
+        "signrawtransactionwithkey",
+        "signrawtransactionwithwallet",
+        "dumpprivkey",
+        "dumpwallet",
+        "importprivkey",
+        "importwallet",
+        "importmulti",
+        "importdescriptors",
+        "sethdseed",
+        "walletpassphrase",
+        "walletpassphrasechange",
+        "encryptwallet",
+    ] {
+        let error = handler
+            .dispatch(method, &json!([]))
+            .err()
+            .unwrap_or_else(|| panic!("{method} unexpectedly succeeded"));
+        assert_eq!(error.code(), RpcError::METHOD_NOT_FOUND);
+        assert!(
+            matches!(&error, RpcError::MethodNotFound(name) if name == method),
+            "{method} must map to MethodNotFound, got {error:?}"
+        );
+    }
     Ok(())
 }
